Simplify steps required to include Python libraries into the IOC build mechanism
A configure/CONFIG.PyDevice is generated when PyDevice is built, and includes
libraries required to link the IOC binary. Since these are pulled from the
same Python environment that PyDevice is built with, they ensure that the IOC will
use the same ones.

A side effect of this patch is that building the IOC no longer needs to execute
python-config. At least on RedHat this means that pythonXX-devel package
needs not to be installed when compiling the IOC. Usually software is built
on the same machine, but it may matter in some cases.

Fixes #11