waveform with string: handling transmitted data
Added a function handles an array of strings and converts
it a string to be processed by python. Uses
Utils.escape to escape the strings.

Processing strings of waveform uses rec->nord instead of
rec->nelm to define its size.

A test for processing was added pydevtest.

pydevtest: PyDev:Array:Pow2 and PyDev:Equal are now
           python3 compatible